+++ src/formatters/scenarioFormatter.ts
@@ -13,13 +13,25 @@ export function format(scenario: Scenario, options?: Partial<FormatOptions>): st
         throw new Error("Scenario must be set!");
     }
     const l = lines(`${scenario.keyword}: ${scenario.name}`);
+    if (scenario.preceedingComment) {
+        l.prepend(scenario.preceedingComment.text);
+    }
     if (scenario.tags.length > 0) {
         l.prepend(formatTag(scenario.tags, options));
     }
+    if (scenario.tagComment) {
+        l.prepend(scenario.tagComment.text);
+    }
     if (scenario.description) {
-        l.append(lines({ trimLeft: true }, scenario.description), null);
+        l.append(lines({ trimLeft: true }, scenario.description));
+    }
+    if (scenario.descriptionComment) {
+        l.append(null, lines(scenario.descriptionComment.text));
     }
     if (scenario.steps.length > 0) {
+        if (scenario.description || scenario.descriptionComment) {
+            l.append(null);
+        }
         const addGroups = config(options).separateStepGroups;
         if (addGroups) {
             scenario.useReadableStepKeywords();