+++ src/formatters/tagFormatter.ts
@@ -11,5 +11,23 @@ export function format(tags: Tag[], options?: Partial<FormatOptions>): string {
         throw new Error("Tags must be set!");
     }
     options = config(options);
-    return lines({ eol: options.oneTagPerLine ? null : " ", indent: "" }, ...tags.map(tag => tag.toString())).toString();
+    if (options.oneTagPerLine) {
+        const l = lines();
+        for (const tag of tags) {
+            if (tag.comment) {
+                l.append(tag.comment.text);
+            }
+            l.append(tag.toString());
+        }
+        return l.toString();
+    }
+    let ls: string = "";
+    for (const tag of tags) {
+        if (tag.comment) {
+            ls += `${ls ? '\n' : ''}${tag.comment.text}\n${tag.toString()}`;
+        } else {
+            ls += `${ls ? ' ' : ''}${tag.toString()}`;
+        }
+    }
+    return lines(ls).toString();
 }