+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 1.1.0 - 2022-02-22
+
+### Added
+
+- Added support for comments in general
+- Added support for language comment
+- Added support for media type and back-tick delimiter in docString
+
 ## 1.0.3 - 2022-01-24
 
 Fix for the indentation issue